fix page view events
